requirements.txt
asgiref==3.9.2
Django==4.2.3
djangorestframework==3.16.1
mysqlclient==2.2.7
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
